Fixing update book issues and adding helper function for User ID
